requirements.txt
# To install the dependencies, run the following command: $ pip install -r requirements.txt

numpy==2.1.3
opencv_contrib_python==4.10.0.84
opencv_python_headless==4.10.0.84
Pillow==11.0.0
